Move SongMetadata chart loading functions to a separate file
